app/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pyjwt>=2.8.0
pytest>=7.4.0
